Splitted into cryptic-test module to keep core without dependencies.
